refactor(cli/migrate): remove unnecessary `Answers` type
