+++ packages/remix-dev/cli/migrate/run/transforms/replace-remix-imports/index.ts
@@ -1,4 +1,5 @@
 import { readFile } from "fs/promises";
+import glob from "fast-glob";
 import { join } from "path";
 import type { PackageJson } from "type-fest";
 
@@ -10,12 +11,8 @@ import type { ExtraOptions } from "./jscodeshift-transform";
 
 const transformPath = join(__dirname, "jscodeshift-transform");
 
-export const updateRemixImports: Transform = async ({
-  answers,
-  files,
-  flags,
-}) => {
-  let pkgJsonPath = join(answers.projectDir, "package.json");
+export const updateRemixImports: Transform = async ({ projectDir, flags }) => {
+  let pkgJsonPath = join(projectDir, "package.json");
   let packageJson: PackageJson = JSON.parse(
     await readFile(pkgJsonPath, "utf-8")
   );
@@ -25,6 +22,11 @@ export const updateRemixImports: Transform = async ({
     content: packageJson,
     path: pkgJsonPath,
     runtime: extraOptions.runtime,
+  });
+
+  let files = glob.sync("**/*.+(js|jsx|ts|tsx)", {
+    cwd: projectDir,
+    absolute: true,
   });
 
   return JSCodeshiftTransform<ExtraOptions>({
